requirements.txt
appdirs==1.4.4
astroid==2.8.0
atomicwrites==1.4.0
attrs==19.3.0
Babel==2.6.0
bcrypt==3.1.7
black==21.9b0
blinker==1.4
boto==2.49.0
boto3==1.17.108
botocore==1.20.108
certifi==2019.11.28
chardet==3.0.4
click==8.0.1
colorama==0.4.3
config==0.5.1
cryptography==2.8
distlib==0.3.1
distro==1.4.0
dj-database-url==0.5.0
Django==2.2.12
django-crispy-forms==1.12.0
django-filter==2.4.0
django-rest-framework==0.1.0
django-storages==1.11.1
djangorestframework==3.12.4
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
future==0.18.2
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
ldap3==2.4.1
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
mypy==0.910
mypy-extensions==0.4.3
netifaces==0.10.4
numpy==1.19.2
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pandas==1.1.3
paramiko==2.6.0
passlib==1.7.2
pathspec==0.9.0
pbr==5.5.0
pexpect==4.6.0
Pillow==7.0.0
pipenv==2020.8.13
platformdirs==2.3.0
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8.6
pyasn1==0.4.2
pycairo==1.16.2
Pygments==2.3.1
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
python-dateutil==2.7.3
python-decouple==3.4
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2021.8.28
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
s3transfer==0.4.2
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.3.12
sqlparse==0.2.4
sshtunnel==0.1.4
stevedore==3.2.2
systemd-python==234
toml==0.10.2
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.25.8
virtualenv==20.0.17
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==0.16.1
whitenoise==5.3.0
wrapt==1.12.1
WTForms==2.2.1
zipp==1.0.0
